+++ src/data/monsters/豺狼守望者/skill.lua
@@ -1,0 +1,11 @@
+local mt = require 'skill.core' '錘擊' {
+    damage_ = {30, 50},
+    proc_ = 0,
+    tip_ = "拿起鏈鎚掃過前方120度的空間，對範圍內的單位造成傷害並擊退200碼，若撞到障礙物則暈眩2秒",
+}
+
+local mt = require 'skill.core' '沖擊波' {
+    damage_ = {70, 100},
+    proc_ = 0,
+    tip_ = "射出一道沖擊波，對路徑上的單位造成傷害",
+}